pyproject.toml
[tool.poetry]
name = "updatebot"
version = "0.1.0"
description = "Automation for updating third party libraries for Firefox"
authors = ["Tom Ritter <tom@mozilla.com>"]
license = "MPL-2.0"

[tool.poetry.dependencies]
python = "^3.5"
requests = "^2.23.0"
pymysql = "^0.9.3"
taskcluster = "^29.1.3"
cryptography = "^2.9.2"
codecov = "^2.1.3"
sentry-sdk = "0.10.2"
json-e = "^4.2.0"
python-dateutil = "^2.8.1"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.2"
coverage = "^5.1"
codecov = "^2.0.22"
flake8 = "^3.7.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
